fix(build errors): Ошибки в sidebar и signUp
- В sidebar.tsx добавил "use client"

SigunUp
- useSearchParams() требует обертывания саспенс на серверной стороне,
- Логика SiginUpPage перенесена в widgets/SiginUpContent
