+++ features/auth/ui/Register/InfoForm/InfoForm.module.css
@@ -35,7 +35,7 @@
   margin-bottom: 36px;
 }
 
-.text_form{
+.text_form {
   display: flex;
   flex-direction: column;
   width: 100%;
