+++ widgets/sidebar/sidebar.tsx
@@ -1,4 +1,4 @@
-import React from "react"
+"use client"
 import {
   SearchIcon,
   HomeIcon,
